backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.2.0
pytest-mock==3.14.0
httpx==0.27.0
